requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
easyocr>=1.7.0
openpyxl>=3.1.0
reportlab>=4.0.0
customtkinter>=5.2.0
img2pdf>=0.5.0
